+++ pkg/builtins/drivers.go
@@ -5,7 +5,7 @@ package builtins
 
 import (
 	// Import all built-in drivers to register them in the registry.
-	_ "github.com/lima-vm/lima/pkg/qemu"
-	_ "github.com/lima-vm/lima/pkg/vz"
-	_ "github.com/lima-vm/lima/pkg/wsl2"
+	_ "github.com/lima-vm/lima/pkg/driver/qemu"
+	_ "github.com/lima-vm/lima/pkg/driver/vz"
+	_ "github.com/lima-vm/lima/pkg/driver/wsl2"
 )
