+++ pkg/driver/qemu/entitlementutil/entitlementutil.go
