+++ pkg/driver/qemu/register.go
