+++ pkg/driver/vz/errors_darwin.go
