+++ pkg/driver/wsl2/lima-init.TEMPLATE
