+++ pkg/qemuimgutil/qemuimgutil.go
