+++ pkg/qemuimgutil/qemuimgutil_test.go
